requirements.txt
Django>=5.0
psycopg2-binary
djangorestframework
django-filter>=25.2
django-simple-history
django-import-export
Pillow>=11.0.0
gunicorn
flake8>=6.0.0
openpyxl>=3.0.0 
